refactor: move TabularNav to SharedComponents
