+++ src/components/CurriculumComponents/CurriculumHeader/CurriculumHeader.tsx
@@ -6,7 +6,7 @@ import Flex from "@/components/SharedComponents/Flex";
 import { Heading, Hr, P } from "@/components/SharedComponents/Typography";
 import SubjectIcon from "@/components/SharedComponents/SubjectIcon/SubjectIcon";
 import Breadcrumbs from "@/components/SharedComponents/Breadcrumbs/Breadcrumbs";
-import TabularNav from "@/components/TabularNav/TabularNav";
+import TabularNav from "@/components/SharedComponents/TabularNav";
 import SubjectPhasePicker, {
   SubjectPhasePickerData,
 } from "@/components/SharedComponents/SubjectPhasePicker/SubjectPhasePicker";
